Export variables to config file
* Move server code out of `src` directory
