+++ package.json
@@ -19,7 +19,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "serve": "node -r esm src/server.js"
+    "serve": "node -r esm ./server/index.js"
   },
   "eslintConfig": {
     "extends": [
